Move Database and models to own packages
